dev: changed postcss icon to a more geometric rendition
The existing PostCSS icon was (seemingly) based on the official version of the icon. It didn't really match my personal taste with how it fit in with the other icons in this set, so I made a new version of the icon that is more simple and geometric (and much smaller in size).