fix: src folder icon open state and .git folder icon open state
The svg for the src folder in it's open state was using the same svg as it's closed state. A new svg with the proper colors was made and substituted for the original.

The .git folder used the folder_github for it's closed state, but the generic open state for it's open state. The fix for this is subjective, but I gave it the same open state as the .github file (file_github_open)